fixed test to accommodate long type

+++ src/test/java/hr/bp/adventofcode/day12/SpringRowTest.java
@@ -14,9 +14,9 @@ public class SpringRowTest {
     public void calculateDifferentArrangements_givenInputOne_returnsCorrectResult() {
         SpringRow springRow = new SpringRow("???.###".chars().mapToObj(c -> (char) c).toList(), List.of(1,1,3));
 
-        int expectedDifferentArrangements = 1;
+        long expectedDifferentArrangements = 1;
 
-        int actualDifferentArrangements = springRow.getDifferentArrangements();
+        long actualDifferentArrangements = springRow.getDifferentArrangements();
 
         Assertions.assertEquals(expectedDifferentArrangements, actualDifferentArrangements);
     }
@@ -25,9 +25,9 @@ public class SpringRowTest {
     public void calculateDifferentArrangements_givenInputTwo_returnsCorrectResult() {
         SpringRow springRow = new SpringRow(".??..??...?##.".chars().mapToObj(c -> (char) c).toList(), List.of(1,1,3));
 
-        int expectedDifferentArrangements = 4;
+        long expectedDifferentArrangements = 4;
 
-        int actualDifferentArrangements = springRow.getDifferentArrangements();
+        long actualDifferentArrangements = springRow.getDifferentArrangements();
 
         Assertions.assertEquals(expectedDifferentArrangements, actualDifferentArrangements);
     }
@@ -36,9 +36,9 @@ public class SpringRowTest {
     public void calculateDifferentArrangements_givenInputThree_returnsCorrectResult() {
         SpringRow springRow = new SpringRow("?#?#?#?#?#?#?#?".chars().mapToObj(c -> (char) c).toList(), List.of(1,3,1,6));
 
-        int expectedDifferentArrangements = 1;
+        long expectedDifferentArrangements = 1;
 
-        int actualDifferentArrangements = springRow.getDifferentArrangements();
+        long actualDifferentArrangements = springRow.getDifferentArrangements();
 
         Assertions.assertEquals(expectedDifferentArrangements, actualDifferentArrangements);
     }
@@ -48,9 +48,9 @@ public class SpringRowTest {
     public void calculateDifferentArrangements_givenInputFour_returnsCorrectResult() {
         SpringRow springRow = new SpringRow("????.#...#...".chars().mapToObj(c -> (char) c).toList(), List.of(4,1,1));
 
-        int expectedDifferentArrangements = 1;
+        long expectedDifferentArrangements = 1;
 
-        int actualDifferentArrangements = springRow.getDifferentArrangements();
+        long actualDifferentArrangements = springRow.getDifferentArrangements();
 
         Assertions.assertEquals(expectedDifferentArrangements, actualDifferentArrangements);
     }
@@ -61,9 +61,9 @@ public class SpringRowTest {
     public void calculateDifferentArrangements_givenInputFive_returnsCorrectResult() {
         SpringRow springRow = new SpringRow("????.######..#####.".chars().mapToObj(c -> (char) c).toList(), List.of(1,6,5));
 
-        int expectedDifferentArrangements = 4;
+        long expectedDifferentArrangements = 4;
 
-        int actualDifferentArrangements = springRow.getDifferentArrangements();
+        long actualDifferentArrangements = springRow.getDifferentArrangements();
 
         Assertions.assertEquals(expectedDifferentArrangements, actualDifferentArrangements);
     }
@@ -72,9 +72,9 @@ public class SpringRowTest {
     public void calculateDifferentArrangements_givenInputSix_returnsCorrectResult() {
         SpringRow springRow = new SpringRow("?###????????".chars().mapToObj(c -> (char) c).toList(), List.of(3,2,1));
 
-        int expectedDifferentArrangements = 10;
+        long expectedDifferentArrangements = 10;
 
-        int actualDifferentArrangements = springRow.getDifferentArrangements();
+        long actualDifferentArrangements = springRow.getDifferentArrangements();
 
         Assertions.assertEquals(expectedDifferentArrangements, actualDifferentArrangements);
     }
